create separate MicrometerMeasurement class and delete the new method then add the arg
