+++ mantis-common/src/main/java/io/mantisrx/common/metrics/measurement/MicrometerMeasurement.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2023 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.mantisrx.common.metrics.measurement;
+
+import io.micrometer.core.instrument.Meter;
+
+public class MicrometerMeasurement {
+        Meter.Type type;
+        double value;
+
+        public MicrometerMeasurement(Meter.Type type, double value) {
+            this.type = type;
+            this.value = value;
+        }
+
+        public Meter.Type getType() {
+            return type;
+        }
+
+        public double getValue() {
+            return value;
+        }
+
+        @Override
+        public String toString() {
+            return "MicrometerMeasurement{" +
+                    "type=" + type +
+                    ", value=" + value +
+                    '}';
+    }
+}
